create event

+++ src/main/java/com/eventtest/repository/UserRepository.java
@@ -1,15 +1,11 @@
 package com.eventtest.repository;
 
 import com.eventtest.model.User;
-import com.eventtest.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import org.springframework.data.jpa.repository.Query;
 
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-//    @Query("SELECT u FROM User u WHERE u.email = ?1")
+public interface UserRepository extends JpaRepository<User, Long> {
+    // @Query("SELECT u FROM User u WHERE u.email =:email")
     User findByEmail(String email);
-
+   // User findByName(String name);
 }
